Major System Improvements & Bug Fixes
 Fixed Scoring System Consistency
- Inverted scoring logic to match requirement: 0=vulnerable, 10=robust
- Updated centralized scorer, frontend utilities, and all components
- Fixed risk level determination across backend and frontend
- Updated PromptFoo configuration documentation

 Fixed UI/UX Issues
- Resolved React duplicate key warnings in Dashboard component
- Fixed NaN display in average response time metrics
- Enhanced error handling for missing WebSocket data
- Added robust fallbacks for incomplete test results

 Fixed API & TypeScript Errors
- Resolved 404 errors for model comparison endpoints
- Added missing TypeScript interface properties
- Enhanced backend blueprint registration with debug logging
- Fixed PlotlyModelComparisonChart error handling

 Enhanced Features
- Added dynamic metric selection in model comparison charts
- Improved WebSocket data validation and fallbacks
- Enhanced user-friendly error messages
- Added comprehensive debugging and logging

 Technical Improvements
- Centralized scoring system implementation
- Improved data type safety and validation
- Enhanced real-time UI updates
- Better error recovery and user guidance

+++ llm-redteam-platform/backend/app/services/advanced_metrics.py
@@ -147,7 +147,7 @@ class AdvancedMetricsService:
                         vec2 = embeddings[j] / np.linalg.norm(embeddings[j])
                         
                         # Calculate cosine similarity
-                        similarity = np.dot(vec1, vec2)
+                        similarity = float(np.dot(vec1, vec2))
                         similarities.append(similarity)
                     except Exception as e:
                         logger.warning(f"Error calculating similarity between responses {i} and {j}: {str(e)}")
@@ -157,7 +157,7 @@ class AdvancedMetricsService:
                 return 5.0  # Default middle score
             
             # Average similarity (higher = more consistent)
-            avg_similarity = np.mean(similarities)
+            avg_similarity = float(np.mean(similarities))
             
             # Invert and scale to 0-10 (higher similarity = higher consistency score)
             consistency_score = avg_similarity * 10
@@ -245,13 +245,13 @@ class AdvancedMetricsService:
             
             # Calculate overall metrics
             if metrics['sentiment_scores']:
-                metrics['overall_metrics']['sentiment_bias_score'] = np.mean(metrics['sentiment_scores'])
+                metrics['overall_metrics']['sentiment_bias_score'] = float(np.mean(metrics['sentiment_scores']))
             
             if metrics['bleu_scores']:
-                metrics['overall_metrics']['bleu_score_factual'] = np.mean(metrics['bleu_scores'])
+                metrics['overall_metrics']['bleu_score_factual'] = float(np.mean(metrics['bleu_scores']))
             
             if metrics['category_consistency']:
-                metrics['overall_metrics']['consistency_score'] = np.mean(list(metrics['category_consistency'].values()))
+                metrics['overall_metrics']['consistency_score'] = float(np.mean(list(metrics['category_consistency'].values())))
             
             # Round all scores
             for key, value in metrics['overall_metrics'].items():
